requirements.txt
streamlit
yfinance
pandas
numpy
plotly
requests
google-genai